rename to `loops` (plural)
