+++ rasa/core/actions/two_stage_fallback.py
@@ -10,7 +10,7 @@ from rasa.core.actions.action import (
     ACTION_DEFAULT_FALLBACK_NAME,
     ACTION_DEFAULT_ASK_REPHRASE_NAME,
 )
-from rasa.core.actions.loop import LoopAction
+from rasa.core.actions.loops import LoopAction
 from rasa.core.channels import OutputChannel
 from rasa.core.constants import USER_INTENT_OUT_OF_SCOPE
 from rasa.core.domain import Domain
